refactor(frontend): restructure some folders and rename filesand  clean up code in sign-up.tsx for better organization and scalability
