+++ client/src/pages/auth/sign-up.tsx
@@ -19,22 +19,22 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { Separator } from "@/components/ui/separator";
-import GoogleOauthButton from "../auth/google-oauth-button";
-
-// Form schema validation
-const formSchema = z
-  .object({
-    fullName: z.string().min(2, "Name must be at least 2 characters"),
-    email: z.string().email("Invalid email address"),
-    password: z.string().min(8, "Password must be at least 8 characters"),
-    confirmPassword: z.string(),
-  })
-  .refine((data) => data.password === data.confirmPassword, {
-    message: "Passwords don't match",
-    path: ["confirmPassword"],
-  });
+import GoogleOauthButton from "../../components/auth/google-oauth-button";
 
 export function SignUp() {
+  // Form schema validation
+  const formSchema = z
+    .object({
+      fullName: z.string().min(2, "Name must be at least 2 characters"),
+      email: z.string().email("Invalid email address"),
+      password: z.string().min(8, "Password must be at least 8 characters"),
+      confirmPassword: z.string(),
+    })
+    .refine((data) => data.password === data.confirmPassword, {
+      message: "Passwords don't match",
+      path: ["confirmPassword"],
+    });
+
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
@@ -49,9 +49,7 @@ export function SignUp() {
     console.log(values);
     // Handle registration logic here
   }
-  const handleClick = () => {
-    window.location.href = `http://localhost:5000/api/auth/google`;
-  };
+
   return (
     <div className="min-h-screen flex items-center justify-center bg-background p-4">
       <Card className="w-full max-w-md mx-auto shadow-lg">
